Fixed multitasking bug Improved page fault handler Added ability to track processes children Raised process limit (may be raised again)

+++ kernel/inc/string.h
@@ -27,6 +27,6 @@ int wcslen(const short *str);
 
  void *memcpy(void *dest, const void *src, u32 n);
 
- void *memset(void *s, u8 c, u32 n);
+ void *memset(void *s, int c, u32 n);
 
 #endif
